+++ Sandboxie/core/dll/net.c
@@ -751,9 +751,9 @@ _FX int WSA_AcceptEx(
     LPOVERLAPPED lpOverlapped)
 {
     //
-    // this call can operate asynchroniusly, hence we can not simply filter here the incoming connection
+    // this call can operate asynchronously, hence we can not simply filter here the incoming connection
     // as we have a proper WFP filter in the driver for now this usermode filtering implementation 
-    // will not filter incoming traffic at all, normally users ate behind NATs or other firewall so 
+    // will not filter incoming traffic at all, normally users are behind NATs or other firewall so 
     // blocking only outgoing connections should be good enough
     //