+++ CHANGELOG.md
@@ -7,7 +7,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [1.6.1b / 5.61.1] - 2022-12-??
 
 ### Fixed
-- fixed support settign page not showing version updates proeprly
+- fixed support setting page not showing version updates proeprly