+++ CHANGELOG.md
@@ -27,7 +27,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [1.6.1a / 5.61.1] - 2022-12-07
 
 ### Added
-- added run boxed command to the sandbox menu
+- added "Run Sandboxed" command to the sandbox menu
 - DefaultBox can now be replaced with another sandbox [#2445](https://github.com/sandboxie-plus/Sandboxie/issues/2445)
 
 ### Fixed